requirements.txt
skia-python==138.0
glfw==2.9.0
PyOpenGL==3.1.9